Adds "upgrade" boxes for 0.12.1
Includes both Trusty & Xenial boxes. Note that the documented procedure of
running `git checkout 0.12.1` was *not* performed here, as the logic
for building multiple platform boxes was not included in the point release
(it had previously been merged into `develop`).

So, I manually set `molecule/shared/stable.ver` to "0.12.1" and ran the logic
from the develop branch. Given that the Ansible changes were included in the point
release, these boxes should still serve as accurate testing for future upgrades.