Runs fetch-tor-packages with pinned version in CI
Adds a `tor_version` var to the fetch-tor-packages logic.
This is yet another point of update when we bump versions,
but pinning will help us avoid surprises. To that end, we're
also running the fetch action in CI now, to help us catch
version mismatches earlier.

Using a separate CI job so as not to add to the serial execution
of the large staging job.