python 2/3 compatibility: revert when python 3 is ready for prod
this commit reverts changes in python->python3, which can only
be deployed when python3 changes are ready for prod. For now, we
plan on having both python2 and python3 compatibility, and we will
update the shebangs to point to /usr/bin/env python3 when we
remove support for python2 and deploy python3 to prod (to be done
in a single release).