More detailed guide to investigating and enabling logs
"Look at these logs" is one of the most common support responses,
so having canonical and correct information in the public docs
is important and  reduces the risk of small errors in these
communications.

I moved the `securedrop-admin logs` instructions into an include
file, since it is a common operation we will likely want to refer
to again in future, and renamed the section to "Investigating
Logs", since it is now hopefully beginning to be deserving of
such a title. ;)

Resolves #4356