Use updated links from Tails website
I left the later link about persistence which describes the things you can enable.
In the list, I used the new link which gives clear, pretty instructions
`Download and verify the Tails .iso and install onto a USB stick or SD card` isn't showing up as a link in the preview, and I don't know why.