Update messages for manually flagged sources for accuracy
The key issue here is that the previous message told sources that their
submission had only just been passed through, which is not true.
This message is also significantly shorter. Removed the icon, as the
"relieved face" emoji, which is only used for this message, is no
longer a good fit with the revised text.

Resolves #4420