Fix Marionette, low-entropy and mock-induced test flakes.
In tests/functional/functional_test.py, FunctionalTest.setup mocked
things up, and if an exception were thrown after that in setup, the
mocks weren't getting cleaned up. Subsequent two-factor tests were
then failing because models.Journalist.verify_token would always
return true. This replaces setup/teardown with pytest fixtures and
ensures that the mocks are stopped.

Also retry web driver creation to work around intermittent Marionette
problems.

We had redundant session timeout tests, so I removed one.

In a couple of places we were changing models.LOGIN_HARDENING and not
resetting it to the original value, so those have been fixed.

Also, because we just set the servers' names to "localhost" in the
tests, Flask was warning that it wasn't a valid cookie domain. So I've
appended ".localdomain" to shut it up. We have enough noise in our
test output.

In test_alembic, the whitespace pattern '\s*' could be empty,
potentially causing trouble for split(). Make it r'\s+'.

Patch get_entropy_estimate in tests/test_integration.py and
tests/test_source.py to prevent failures when CI machines run low on
entropy. In test_source.py, the call to logger.assert_called_once_with
in test_failed_normalize_timestamps_logs_warning fails if entropy is
low in CI -- it's failing not because the message is repeated, but
because the mock is called more than once, when the submit route logs
that it can't generate the source's key.