Lint/CI fixes
Make sure lint targets activate the virtualenv as necessary to pick up
pytest, since not everyone will have it somewhere else on the
path. Ahem.

Normalize lint target between Makefiles. Rename some variables in the
top-level Makefile.

Circle's still running the EOL Ubuntu 14.04, which has Python 3.4. We
can run most linters under 3.5 in our 16.04 container, but the
shellcheck package there is too old, so unless we want to run
Docker-in-Docker, we need to run shellcheck in a separate step using a
koalaman image.