Replace "Cloud" upload icon with more neutral one
Adds rendering rule for the icon to accomplish two things:
- give the icon room to breathe
- force crisp rendering, to render large source image cleanly
  at smaller size (the source PNG is a bit larger than the
  default size to accommodate zooming)

The crisp rendering rule may make sense as a generic rule at a
later point, but so far this seems to be the best candidate
for using it.

The new icon was designed by Nina Alter (@ninavizz) and is not
derived from FontAwesome, hence not in the font-awesome directory.

Resolves #4535