Support simultaneous dev containers
Teach dev-shell to use a prefix for ports exposed in the dev
containers, when necessary. This permits running multiple
container-based tasks (anything run with dev-shell in the Makefile)
simultaneously.

For example, if you run "make dev" with no other containers running,
VNC will be on 5909, the source interface on 8080, and the journalist
interface on 8081. If you then run "make test" while that container is
running, the test container's services should be available at 15909,
18080, and 18081. Obviously the number of containers we can run with
this simple scheme is limited, but probably sufficient.