Improve deletion of submissions
Replace srm with shred, which is faster, reducing the chance that
deletion of a submission will be interrupted.

Update rq and redis requirements, to eliminate long-standing bugs.

Add manage.py tasks for detecting and correcting submissions that have
been disconnected from their files on disk, and vice versa. Update
manage.py to explicitly run with the production virtualenv. Also
specify the virtualenv in WSGI scripts and the run-test script. In the
dev/test Docker container, install requirements in a virtualenv at the
same path as production.

Add a supervisor script for requeuing interrupted rq jobs. If the app
server is rebooted while an rq job is running, that job has already
been deleted from the queue and rq will not automatically resume it on
reboot, but it does have a record of it in the queue's started job
registry. This script checks that registry for jobs that aren't
already queued or being run, and requeues them.