Overhaul Transfer Device and export recommendations
Resolves #4620
Resolves #4646
Resolves #4434
Resolves #4670

In addition to introducing the Transfer and Export Device,
this commit clearly breaks out optional hardware into its
own checklist, more strongly recommends purchase of a
printer, and strengthens recommendations for malware
mitigation.

It removes some outdated recommendations and a reference to
storing the journalist's GPG passphrase in KeePassX
(not mentioned anywhere else).

It removes instructions for journalists to set up individual
GPG keys, as they are unlikely to be followed, and the more
critical recommendation is to ensure tha the Export Device
is encrypted.

It updates the overview diagram and data flow diagram
consistent with the above changes. This update also makes
the diagrams more consistent with terminology and current
practices. It removes OnionShare from the data flow
diagram, as it is not currently mentioned elsewhere in the
docs.

(cherry picked from commit 948039b50d679b87b32f354ddbe95f863c5bd9cc)