alembic: migration contains raw SQL, add `# nosec`
bandit flags this because the table name is passed in as a variable.
It's not user controlled so this does not introduce a security problem
(function is in the migration only). I could suppress the
alert by removing the function but in the spirit of not
making this alembic migration even more repetitive I'm adding `# nosec`.