Updates fail-fast logic for v2 -> v3 transition
Checks that BOTH of the following conditions are NOT true:

  * v3_onion_services=false
  * v3 services exist on remote host

If both are true, that's likely the result of a misconfiguration.
This logic handles the prod use case of v2 enabled, also enabling
v3. It hasn't yet been tested with ssh-over-LAN, we'll likely need to
modify the restrict-direct-access role to accommodate.