securedrop-config: remove function setting HiddenServiceVersion
For new installs that wish to use v2, the HiddenServiceVersion
will be set via the torrc template in the tor-hidden-services
Ansible role.

For existing installs that still wish to use v2, if they have
updated to Xenial, then they already have this HiddenServiceVersion
set to 2 in their existing torrc. Installs _not_ on Xenial
are no longer supported.