Manage rq services with systemd instead of supervisor
This change makes securedrop-app-code Conflict/Replace supervisor, to
ensure that supervisor is purged from the system. The rq worker and
requeuing processes are now managed as systemd services, with
configuration files in /lib/systemd/system.

Submission hashing and deletion works as before. Requeuing a job
interrupted by a reboot is less likely to work, as the worker seems to
be shut down such that the running jobs are terminated and moved by rq
to the failed queue. Files surviving that will be reported in the next
daily check, and can be deleted with manage.py.

The services now log to the systemd journal, which in our current
configuration is stored only in memory, so will not survive
reboots. This could be changed by creating /var/log/journal, but if we
do that we'll need to audit what other systemd services are logging.