Fixes v3 keypair generation in upgrade box prep
During 1.0, we did not test the v3 keypair generation in the upgrade
scenario. Makes sense: we only run that scenario *after* a release is
final. The git-repo-root logic assumed we were in Tails or running
against normal staging VMs. In the vagrant-package logic (which builds
"upgrade" boxes for testing), the proper path structure requires that
filepaths be relative to the secondary git repo, cloned in order to
prepare an old version from the known-good tag.

This var override ensures that the v3 keys are stored inside the
secondary git repo. We could instead update the rev-parse logic to be
relative to the role, but that'd require additional testing in Tails,
and this works just as well.