Fix column/button alignment problems on source start page
Certain combinations of language and viewport size could result in the
two columns on the source start page not lining up with each
other. This changes the columns' justify-content property to push the
top and bottom containers apart, and controls the borders and content
of the buttons to align them better.