Fix Tails detection in Python 3 + Tails 3.x
The final comparison failed to account for the fact that the
subprocess module returns a byte literal.

Incorrect Tails detection caused torify not to be used, which
caused a network error during the creation of the virtualenv
on Tails 3.x.

Decoding the string first reduces the likelihood of such
coding errors.

Fixes #4925