Fix "make test-config"
Outside of a dev container, "make test-config" was broken since the
shebang line in i18n_tool.py was pointed to
/opt/venvs/securedrop-app-code. When run with dev-shell, the generated
config was broken by all the Docker build noise. This change updates
dev-shell to observe the DOCKER_BUILD_VERBOSE environment variable and
silence the docker build if it's not "true", and generates the test
config with dev-shell, with DOCKER_BUILD_VERBOSE=false.