Make clean action now supports staging scenario
We'd neglected to add support for the Molecule-based staging env to the
"make clean" action. Refactored the associated scripts a bit, to enable
emitting just the scenario name, for use in different contexts (e.g.
create, destroy).

This solution isn't perfect. In the following edge case:

1. `make staging`
2. reboot host machine
3. `make clean` # will fail

The problem is that Molecule stores VM state in `/tmp/`, which is wiped
on reboot. So Molecule will determine that the VMs don't exist, but
libvirt will (rightly!) disagree. The solution in that case is to run
`virt-manager` and clean up the VMs manually. We *could* script such via
`sudo virsh`, but I'm loath to place sudoized commands in scripts to run
on developer machines without confirmation.