Prevent duplicate instances of source /generate page (#4458)
If a source clicks on 'Get Started' in two separate browser windows or tabs
the second instance will redirect the source to the source index page with
a flash message stating that another window is already on the 'Get Started'
page. This addresses #4458 by ensuring only one instance of the /generate
page can be opened at once.