For secure deletion, use a systemd service instead of rq
Instead of submitting jobs to rq that invoke rm.secure_delete, move
files to be deleted to a designated directory in the store, and use a
systemd service, securedrop_shredder, to monitor that directory and
clear its contents securely.

Moving files is atomic and quick, so should never be interrupted. If
the secure deletion is interrupted, the securedrop_shredder service
will just find the file on its next pass.

We had discussed adding columns to the database tables involved, and
having the service use those to discover items to delete, but:

- It would have been a bigger, more complicated change.

- It would have introduced the possibility of source information being
  present in the database for quite a while after the admin "deleted"
  it, which could be problematic in the event of raid/seizure. This
  way only the encrypted files might linger; the database records will
  still be deleted immediately.

- It would have required that we always remember to filter using those
  columns to avoid working with deleted content. (SQLAlchemy currently
  has only partial support for setting up this filtering globally:
  https://github.com/sqlalchemy/sqlalchemy/wiki/FilteredQuery)