Handle multiple instances of source /generate (#4458)
When the source opens more than one browser tabs on /generate before clicking
on SUBMIT DOCUMENTS:
-Flashes a message promting the source to verify their codename in /lookup
 after they click on SUBMIT DOCUMENTS in a fist instance of /generate.
-Redirects the source to /lookup when they click on SUBMIT DOCUMENTS in a
 second instance of /generate and displays a flash message prompting the
 source to verify their codename.
-Makes the codename in /lookup visible by default.

This eliminates the 500 server error reported in #4458 and reduces the
potential of codename confusion following the generation of multiple
codenames in different browser tabs or windows.

The behavior when only one instance of /generate is opened remains unchanged.