Parallelize translation tests in CI
Split the CI translation test job up by locale instead of test file,
and increase the parallelism to 20. This lets the current set of 20
locales finish in less than 20 minutes; as we add languages we may
want to increase the container count further, or we may decide that
other work needs to get through CI too and the LM shouldn't be such a
pig.

To make the parsing and splitting of our supported locales a little
nicer, I added options to the list-locales function of
i18n_tool.py. Now the default is to print a space-separated list
suitable for most shell use; --python prints a Python list suitable
for inclusion in the SecureDrop config.py; and --lines prints one
locale per line, suitable for feeding to `circleci tests split`, among
other things.