Handle multiple instances of /generate
Associate codenames from different browser tabs open to /generate with
a unique tab id hidden in the form. This allows to register the source
using the codename displayed in the browser tab from which the source
proceeds to submit documents instead of the last codename generated
(current behavior).

Proceeding to submit documents from another tab open to /generate
redirects to /lookout with a flash message informing the source they are
already logged-in and to verify their codename. The codename is made
visible at the bottom of the /lookup page.