Try to fix CI Docker image caching without ditching path hash
The hash of the repo path enables -- or at least was _intended_ to
enable -- working with two local checkouts, even if the changes
involved the Dockerfile. Unfortunately it wasn't used everywhere it
needed to be, like, oh, in the `docker run` command line. This
corrects that and also uses the path hash in the CircleCI image
naming.