Pins pip, setuptools, pip-tools
The goal is to provide recent and predictable versions of pip and
setuptools. Also upgraded the version of pip-tools without it,

Adds --allow-unsafe in app-code deps

The pip-compile command *must* use --allow-unsafe in order to pin
seutptools and/or pip. We now do so. Adding this flag worked fine during
local invocation, i.e. running `pip-compile` on the host machine
(plus a few apt deps such as apache2-dev, enchant, etc),
but fails via the containerized make target with:

Updates all the different requirements.txt file to pull in the latest changes.