Clarify ordering of a few operations in release management docs
The old instruction to "Prepare a changelog describing the changes in
the release" sounded like you should be updating the changelog files
at that point, when that happens automatically during the execution of
"update_version.sh".

Also improved some indentation to clarify flow, and switched to
auto-enumerated lists to avoid having to renumber when steps are added
or deleted.

For upgrade testing, removed the suggestion to build upgrade boxes on
develop for point releases.