Improve admin workstation updater GUI error handling
Increase the time it will wait for tailsconfig to run.

Don't say the admin password was wrong on tailsconfig
timeout. Instead, look for sudo in the output to determine if an
incorrect password was given.

Add a new message explaining that tailsconfig took too long, and don't
say "Exiting upgrade" when an incorrect admin password was given, as
the updater is not exiting, and in fact we're telling the admin to try
again.

Make sure the failure reason was cleared on each attempt.