Fix tor restart handler naming
Apparently Ansible used to inexact matching when looking for
handlers. That changed in the 2.8 series. See:

  https://github.com/ansible/ansible/issues/55575

That broke the fuzzy name matching relied upon in the
restart-tor-naming handler, causing Tor not to be restarted
properly. We could also specify a consistent topic to monitor with
"listen", but I think this restart logic can be simplified to one task
anyway.