Changes include -> include_tasks in tor handler
The Ansible 2.8 porting guide [0] states that handlers cannot use
"include", they should instead use "include_tasks". Done. In the
"upgrade" scenario, however, the old version of the Ansible configs are
cloned in via git, and those will still not be compatible with Ansible
2.9, so let's patch the handler file in place as a shim to provide
mutual compatibility with the old & new Ansible versions.

The patch task should be conditional on a semvar comparison, rather than
a substring match, but in practice we only ever use the upgrade scenario
to test adjacent versions anyway, so didn't seem worth the trouble.

This commit also renames a handle for uniqueness. That wasn't strictly
required for compatibility, but it makes it easier to discover run order
when reading or grepping the config.

[0] https://docs.ansible.com/ansible/latest/porting_guides/porting_guide_2.8.html#imports-as-handlers