Use WTForm for source interface submission form
Adds SubmissionForm to securedrop/source_app/forms.py, supporting
validation of the source submission form. Rework lookup.html to use
it: some presentation (the placeholder) had to be moved to the form,
because the translated message couldn't be interpolated into the
SubmissionForm construction in the template.

Note that this change does *not* add the maxlength attribute to the
message textarea, as Firefox doesn't limit the text length accurately
when it includes newlines and can submit a message longer than the
validator limit.

(cherry picked from commit 412103d74b52f8ea8f3e110cb3f55135372ab500)