Remove unused/disabled source onion service info files
If a site disables v2 onion services, remove the file used to decide
to show the deprecation warning.

The v3 equivalent is also removed if v3 services are disabled, though
downgrading to v2-only is not really possible -- more work would be
needed in the playbooks, and that's probably not something we should
support at this point.

(cherry picked from commit 492db2deb7b0667af9f3c3b646785807400fddbc)