Get CI parallelization of translation tests working again
The out-of-date Ubuntu image couldn't get through an apt update
because of the Google Chrome repo, so python-sh couldn't get
installed, which broke the invocation of i18n_tool.py to list locales
for CircleCI splitting.

This updates the machine image, and also improves the locale argument
handling in the translation-test Makefile target.