Update dev container browser software and validation
Mozilla's keyserver went down, and they might just leave it that
way. Their release signing key is available on another one, but that
doesn't seem to be the result of any official policy I can find. The
closest thing to guidance comes from a blog post instructing us to
obtain the key from the archive server along with the
software[1]. This changes the Dockerfile to do that.

Recent versions of geckodriver are being signed with the release key,
so we'll now use it to verify that package too.

We were also using an expired version of the Tor release signing key,
kept in our repo, so I'm following their instructions[2] for obtaining
the key instead.

Finally, this updates all three packages.

[1] https://blog.mozilla.org/security/2019/06/13/updated-firefox-gpg-key/
[2] https://support.torproject.org/tbb/how-to-verify-signature/