Fixes Xenial docker build (#5477)
* Fixes xenial docker build

We need a recent version of dh-virtualenv. As for Focal, for Xenial we
now pull from the Debian [sic] unstable [sic] repos to get a recent
version. In Focal, we could reuse the already-available
debian-archive-keyring to verify integrity, but in Xenial, the
debian-archive-keyring is too old, Jessie-era (2014-11 vs 2018-09,
judging by the timestamps). So for Xenial, we must fetch the required
key over HTTPS and provide that to apt manually.

We must also explicitly reference the "unstable" target for
dh-virtualenv, since an older version of the package is available in
Xenial, and will be preferrred to the unstable version given the
cautious apt preferences we configure.

* Stop dh-virtualenv pulling in Sid Python 3 packages

* Updates xenial builder image hash

Fixes to the build logic were just added by @rmol, so I've rebuilt the
image and pushed it to facilitate testing by others.

* Adds tests to package build logic for py/dhv

We now test explicitly for expected versions of:

  * python3
  * dh-virtualenv

Since we're pulling from non-Ubuntu repositories. We'll reuse
the same test logic in the march toward Focal build support, as well.

Co-authored-by: John Hensley <john@freedom.press>