Update tor upgrade check for newer aptitude
Between the version of aptitude on Xenial and that on Focal, empty
searches started returning 1, causing `Check whether tor will be
upgraded` to fail. Let's ignore the exit code, since what we care
about is whether "tor" is in the stdout.

We could use `apt list --upgradable`, but that command warns about
lack of a stable CLI, and additionally we'd have to munge the output
further to inspect it.