Updates PaX flag header style based on OS
For Xenial we'll continue using the C flag but under Focal we must start
using "c" instead. Provides a map in the Ansible settings so the right
flag is used depending on what OS the servers are running.

Notably we're overlooking the "securedrop-grsec" metapackage which also
has these same flags hardcoded, but we may decide to switch to paxctld
instead.

The reformatting on the "when" conditional non-functional, but silences
an Ansible deprecation warning.