Update bandit to 1.7.0
The 322 test was not yet deprecated in our last pinned version (1.4.0)
so removing it from the skip list in #5670 broke in local environments
in which bandit wasn't upgraded (as is done each run in CI).

Also, 1.4.0 didn't yet support glob patterns in bandit's --exclude
option, so running it locally would take forever as the .venv
directory was scanned.