Update translated strings
Our current i18n release management workflow gets everything right on
the securedrop-i18n/i18n branch, but doesn't ensure that the PO
template is updated back on the develop branch. This commit brings
securedrop/translations/messages.pot up to date. A documentation
update is pending to reduce the chance of this happening in the
future.