Add localization checks to basic source/journalist interface tests
We haven't really been testing translations. In the page layout tests,
the content of strings is often only checked after making sure the
page is *not* translated. This has led to us not catching text which
didn't get marked for translation, thus being shown in English to
users who are using a different locale.

This commit includes the following changes to make it easier to test
that source strings are properly translated:

- a context manager which will xfail a test for a locale if any of the
message IDs passed to it are not present in the locale's message
catalog

- a pytest flaky filter which prevents retries of expected failures

- turning off flaky's reporting of successful tests

- functions to obtain test locales and counts to use in testing their
pluralization rules, which can be used with pytest.parametrize to
easily run a test in multiple locales, including fuzzing plural
strings to check that translations are complete, while only trying as
many variations as the rules need.

Adding locale parameterization to tests increases the overall time
they take, of course: test_journalist.py went from around 2.5 minutes
to 3.5 minutes.