Improve i18n_tool.py update-from-weblate
When committing changes to a .po file, the script assumed that it
would always have committed the last change as part of the
update-from-weblate task. If the file were changed outside of that
workflow, and the commit message wasn't structured as expected, the
script would be unable to gather translator credits correctly.

This changes the script to look past the last commit until it finds
one of its commit messages.