Upgrade boxes for 1.8.0, Xenial (not working)
Rebuilds 1.8.0 Xenial boxes

We've had some trouble running the 1.8.0 upgrade boxes. This rebuilds
them, using the latest available bento/16.04 version (202103.19.0),
as well as a small tweak to the prep logic to provide a depedency that
was recently removed from the Bento sources.

Even with the rebuild, still isn't working. Will pivot and focus on
Focal (ha!) support.