Fix some translation problems
Correct some misplaced parentheses in gettext/format calls.

Mark some strings that were being presented untranslated.

Fix locale/session handling on CSRF error. We were translating the
CSRF error message, then clearing the session and showing the rest of
the page in the default locale, almost certainly English. We need to
reevaluate how we want to handle the locale when clearing the session,
but for now, this just fixes the inconsistency.