Generalize procurement of server OS information
Not all Linux distributions have /etc/lsb-release. Debian 10, for
instance, only has /etc/os-release, which means the server_os module
could be broken in a dev environment outside of containers. This
change uses the lsb_release command to obtain the operating system
version.

It also uses functions instead of module attributes for the OS release
and Xenial EOL date, in an attempt to improve patching in
tests. Manipulating the module attributes in tests was causing
consistent flaky (oh, the irony) retries.