Adds new pubkey for Release Signing Key
Bumps the version of the `securedrop-keyring` package, preserving the
old/current release signing key, but adding a new pubkey.

  * Old/current fingerprint: 22245C81E3BAEB4138B36061310F561200F4AD77
  * New/next fingerprint: 2359E6538C0613E652955E6C188EDD3B7B22E6A3

This this is a soft rotation, we'll make sure that all instances have
the new key first, then later remove reference to the old key.

As with previous updates, here's the command I used to import the new
key into the keyring:

    gpg --no-default-keyring --keyring \
    install_files/securedrop-keyring/etc/apt/trusted.gpg.d/securedrop-keyring.gpg \
    --import \
    install_files/ansible-base/roles/install-fpf-repo/files/fpf-signing-key-2021.pub

Note the tweak to the target file to import, i.e. the "2021" suffix.