Generate source keypairs synchronously without checking entropy pool
The changes in sources of randomness in Linux, and libgcrypt's use of
them, have eliminated any need there might have been to check the
amount of entropy available. Keypair generation takes less than a
second on current CPUs. There's no need to defer key generation or to
perform it asynchronously.