Replace alembic merge with version order tweak.
A merge version is recommended for situations where alembic history
has multiple heads, but in this case that was breaking the upgrade/downgrade
tests. Since the db changes in question are not yet in prod, it's ok
to reorder the history instead - so the otp_secret change has been moved from
its own alembic branch to the head of the existing one.