Added --ignores for Ansible make safety warnings
The ignored safety ids are 42923 and 42926 - both relate to bugs where
async data stored on a remote host may be inadvertantly visible to users
on said remote host. This is not directly applicable to SecureDrop's use of
Ansible - a fix can wait until there is either a 2.9.* version with the
backported change or until the Ansible major version is updated.