third pass through "/col/<filesystem_id>": fixes "col.html" styling
NB. As in "index.html" and "_source_row.html", the conversion of
UL.submissions into TABLE.submissions is not pixel-perfect.  In this
case, I've not even attempted to make it so: this layout is to my eye
more idiomatic visually as well as semantically.  This too is up for
debate.