Support using podman for the dev environment
podman is mostly a drop-in replacement for Docker that's lighter and
better supports rootless execution. It tends to be more widely used
in the Fedora/Red Hat world, though was recently added to Debian. It
does not require running a background daemon as root, providing some
nice security benefits.

By setting a `USE_PODMAN=1` environment variable, the dev-shell and
related scripts will invoke `podman` instead of `docker.

We also need to pass `--userns=keep-id` to `podman run` so that the host
UID/GID are properly mapped in the container so files are writable.

podman doesn't allow deleting/overwriting `/dev/random` so that step is
optional now. I assume that's OK because this is just a dev environment.

This is mostly based on <https://github.com/wikimedia/fresh/commit/d1535438cfd8ccaf39d18fa2fd43350526e43401>,
which I had contributed to/reviewed.

In the future it would be nice to automatically detect if podman is
installed when docker can't be found, but that should probably wait
for more testing/usage under podman.